collector: Simplify collecting routine in single source file
